Django_repos/incident_mgmt/incidents/serializers.py
from rest_framework import serializers
from .models import Incident

class IncidentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Incident
        fields = '__all__'
        read_only_fields = ['incident_id', 'reported_at', 'user', 'reporter_name']
