Django_repos/incident_mgmt/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.20
djangorestframework==3.15.2
geographiclib==2.0
geopy==2.4.1
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
